+++ src/api/network_test.go
@@ -2,6 +2,7 @@ package api
 
 import (
 	"encoding/json"
+	"errors"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
@@ -14,14 +15,42 @@ import (
 )
 
 func TestConnection(t *testing.T) {
+	bp := daemon.BlockchainProgress{
+		Current: 35,
+		Highest: 39,
+		Peers:   nil}
+	bp.Peers = append(bp.Peers, struct {
+		Address string `json:"address"`
+		Height  uint64 `json:"height"`
+	}{
+		Address: "127.3.5.1",
+		Height:  39,
+	})
+	bp.Peers = append(bp.Peers, struct {
+		Address string `json:"address"`
+		Height  uint64 `json:"height"`
+	}{
+		Address: "127.0.0.1",
+		Height:  12,
+	})
+	bp.Peers = append(bp.Peers, struct {
+		Address string `json:"address"`
+		Height  uint64 `json:"height"`
+	}{
+		Address: "127.0.5.1",
+		Height:  13,
+	})
+
 	tt := []struct {
-		name                       string
-		method                     string
-		status                     int
-		err                        string
-		addr                       string
-		gatewayGetConnectionResult *daemon.Connection
-		result                     *daemon.Connection
+		name                               string
+		method                             string
+		status                             int
+		err                                string
+		addr                               string
+		gatewayGetConnectionResult         *daemon.Connection
+		gatewayGetBlockchainProgressResult *daemon.BlockchainProgress
+		gatewayGetBlockchainProgressError  error
+		result                             *daemon.Connection
 	}{
 		{
 			name:   "405",
@@ -44,6 +73,8 @@ func TestConnection(t *testing.T) {
 			status: http.StatusOK,
 			err:    "",
 			addr:   "addr",
+			gatewayGetBlockchainProgressResult: &bp,
+			gatewayGetBlockchainProgressError:  nil,
 			gatewayGetConnectionResult: &daemon.Connection{
 				ID:           1,
 				Addr:         "127.0.0.1",
@@ -65,12 +96,35 @@ func TestConnection(t *testing.T) {
 				ListenPort:   9877,
 			},
 		},
+		{
+			name:   "500 - blockchain progress failed",
+			method: http.MethodGet,
+			status: http.StatusInternalServerError,
+			err:    "500 Internal Server Error - some error",
+			addr:   "addr",
+			gatewayGetBlockchainProgressResult: nil,
+			gatewayGetBlockchainProgressError:  errors.New("some error"),
+			gatewayGetConnectionResult: &daemon.Connection{
+				ID:           1,
+				Addr:         "127.0.0.1",
+				LastSent:     99999,
+				LastReceived: 1111111,
+				Outgoing:     true,
+				Introduced:   true,
+				Mirror:       9876,
+				ListenPort:   9877,
+			},
+		},
 	}
 	for _, tc := range tt {
 		t.Run(tc.name, func(t *testing.T) {
 			endpoint := "/network/connection"
 			gateway := NewGatewayerMock()
 			gateway.On("GetConnection", tc.addr).Return(tc.gatewayGetConnectionResult)
+			gateway.On("GetBlockchainProgress").Return(
+				tc.gatewayGetBlockchainProgressResult,
+				tc.gatewayGetBlockchainProgressError,
+			)
 			v := url.Values{}
 			if tc.addr != "" {
 				v.Add("addr", tc.addr)
@@ -102,13 +156,41 @@ func TestConnection(t *testing.T) {
 }
 
 func TestConnections(t *testing.T) {
+	bp := daemon.BlockchainProgress{
+		Current: 35,
+		Highest: 39,
+		Peers:   nil}
+	bp.Peers = append(bp.Peers, struct {
+		Address string `json:"address"`
+		Height  uint64 `json:"height"`
+	}{
+		Address: "127.3.5.1",
+		Height:  39,
+	})
+	bp.Peers = append(bp.Peers, struct {
+		Address string `json:"address"`
+		Height  uint64 `json:"height"`
+	}{
+		Address: "127.0.0.1",
+		Height:  12,
+	})
+	bp.Peers = append(bp.Peers, struct {
+		Address string `json:"address"`
+		Height  uint64 `json:"height"`
+	}{
+		Address: "127.0.5.1",
+		Height:  13,
+	})
+
 	tt := []struct {
-		name                        string
-		method                      string
-		status                      int
-		err                         string
-		gatewayGetConnectionsResult *daemon.Connections
-		result                      *daemon.Connections
+		name                               string
+		method                             string
+		status                             int
+		err                                string
+		gatewayGetConnectionsResult        *daemon.Connections
+		gatewayGetBlockchainProgressResult *daemon.BlockchainProgress
+		gatewayGetBlockchainProgressError  error
+		result                             *daemon.Connections
 	}{
 		{
 			name:   "405",
@@ -121,6 +203,8 @@ func TestConnections(t *testing.T) {
 			method: http.MethodGet,
 			status: http.StatusOK,
 			err:    "",
+			gatewayGetBlockchainProgressResult: &bp,
+			gatewayGetBlockchainProgressError:  nil,
 			gatewayGetConnectionsResult: &daemon.Connections{
 				Connections: []*daemon.Connection{
 					&daemon.Connection{
@@ -150,12 +234,38 @@ func TestConnections(t *testing.T) {
 				},
 			},
 		},
+		{
+			name:   "500 - blockchain progress failed",
+			method: http.MethodGet,
+			status: http.StatusInternalServerError,
+			err:    "500 Internal Server Error - some error",
+			gatewayGetBlockchainProgressResult: nil,
+			gatewayGetBlockchainProgressError:  errors.New("some error"),
+			gatewayGetConnectionsResult: &daemon.Connections{
+				Connections: []*daemon.Connection{
+					&daemon.Connection{
+						ID:           1,
+						Addr:         "127.0.0.1",
+						LastSent:     99999,
+						LastReceived: 1111111,
+						Outgoing:     true,
+						Introduced:   true,
+						Mirror:       9876,
+						ListenPort:   9877,
+					},
+				},
+			},
+		},
 	}
 	for _, tc := range tt {
 		t.Run(tc.name, func(t *testing.T) {
 			endpoint := "/network/connections"
 			gateway := NewGatewayerMock()
 			gateway.On("GetConnections").Return(tc.gatewayGetConnectionsResult)
+			gateway.On("GetBlockchainProgress").Return(
+				tc.gatewayGetBlockchainProgressResult,
+				tc.gatewayGetBlockchainProgressError,
+			)
 			req, err := http.NewRequest(tc.method, endpoint, nil)
 			require.NoError(t, err)