+++ src/api/wallet.go
@@ -56,6 +56,12 @@ type WalletResponse struct {
 	Entries []WalletEntry `json:"entries"`
 }
 
+// BalanceResponse address balance summary struct
+type BalanceResponse struct {
+	wallet.BalancePair
+	Addresses wallet.AddressBalance `json:"addresses"`
+}
+
 // NewWalletResponse creates WalletResponse struct from *wallet.Wallet
 func NewWalletResponse(w *wallet.Wallet) (*WalletResponse, error) {
 	var wr WalletResponse
@@ -113,7 +119,7 @@ func walletBalanceHandler(gateway Gatewayer) http.HandlerFunc {
 			return
 		}
 
-		b, err := gateway.GetWalletBalance(wltID)
+		walletBalance, addressBalances, err := gateway.GetWalletBalance(wltID)
 		if err != nil {
 			logger.Errorf("Get wallet balance failed: %v", err)
 			switch err {
@@ -129,7 +135,10 @@ func walletBalanceHandler(gateway Gatewayer) http.HandlerFunc {
 			return
 		}
 
-		wh.SendJSONOr500(logger, w, b)
+		wh.SendJSONOr500(logger, w, BalanceResponse{
+			BalancePair: walletBalance,
+			Addresses:   addressBalances,
+		})
 	}
 }
 
@@ -171,6 +180,12 @@ func getBalanceHandler(gateway Gatewayer) http.HandlerFunc {
 			return
 		}
 
+		// create map of address to balance
+		addressBalances := make(wallet.AddressBalance, len(addrs))
+		for idx, addr := range addrs {
+			addressBalances[addr.String()] = bals[idx]
+		}
+
 		var balance wallet.BalancePair
 		for _, bal := range bals {
 			var err error
@@ -187,7 +202,10 @@ func getBalanceHandler(gateway Gatewayer) http.HandlerFunc {
 			}
 		}
 
-		wh.SendJSONOr500(logger, w, balance)
+		wh.SendJSONOr500(logger, w, BalanceResponse{
+			BalancePair: balance,
+			Addresses:   addressBalances,
+		})
 	}
 }
 
@@ -289,7 +307,7 @@ func walletSpendHandler(gateway Gatewayer) http.HandlerFunc {
 		}
 
 		// Get the new wallet balance
-		b, err := gateway.GetWalletBalance(wltID)
+		walletBalance, _, err := gateway.GetWalletBalance(wltID)
 		if err != nil {
 			err = fmt.Errorf("Get wallet balance failed: %v", err)
 			logger.Error(err)
@@ -297,7 +315,7 @@ func walletSpendHandler(gateway Gatewayer) http.HandlerFunc {
 			wh.SendJSONOr500(logger, w, ret)
 			return
 		}
-		ret.Balance = &b
+		ret.Balance = &walletBalance
 
 		wh.SendJSONOr500(logger, w, ret)
 	}